对查询的修正优化，查询的部分测试

+++ liuwwww/liuwwww-jdbc/src/main/java/org/liuwww/db/dao/IQueryDao.java
@@ -22,5 +22,5 @@ public interface IQueryDao
 
     public Page getPage(SqlBean bean, Page pageInfo) throws SysException;
 
-    public int getCount(SqlBean bean) throws SysException;
+    public long getCount(SqlBean bean) throws SysException;
 }
