+++ liuwwww/liuwwww-jdbc/src/main/java/org/liuwww/db/query/QueryBean.java
@@ -44,7 +44,7 @@ public interface QueryBean
 
     public QueryBeanCompare getCompare();
 
-    public int getCount();
+    public long getCount();
 
     public SqlBean getSqlBean();
 }
